backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
torch==2.1.1
boto3==1.33.1
numpy==1.24.3
python-dotenv==1.0.0
